+++ sdk/servicebus/service-bus/samples/v7/typescript/src/deleteMessages.ts
@@ -53,7 +53,7 @@ export async function main(): Promise<void> {
     peekedMessages = await queueReceiver.peekMessages(max32BitNumber);
     console.log(`Peeked messages (1): ${peekedMessages.length}.`); // should be 10
 
-    let { deletedCount } = await queueReceiver.deleteMessages({ maxMessageCount: 10 });
+    let { deletedCount } = await queueReceiver.deleteMessages(10);
 
     console.log(`Number of messages deleted: ${deletedCount}.`);
 
@@ -67,8 +67,7 @@ export async function main(): Promise<void> {
     peekedMessages = await queueReceiver.peekMessages(max32BitNumber);
     console.log(`Peeked messages (2): ${peekedMessages.length}.`); // should be 20
 
-    ({ deletedCount } = await queueReceiver.deleteMessages({
-      maxMessageCount: 20,
+    ({ deletedCount } = await queueReceiver.deleteMessages(20, {
       beforeEnqueueTime: timeMarkUtc,
     }));
     console.log(`Number of messages deleted this time: ${deletedCount}.`); // should be 10